+++ libs/globetrotter/util/src/index.ts
@@ -11,6 +11,6 @@ export * from './lib/domain/summary.interface';
 
 export * from './lib/duration';
 export * from './lib/quiz-type.enum';
-export * from './lib/route.enum';
+export * from './lib/routes';
 export * from './lib/selection.interface';
 export * from './lib/shuffle';
